camera_tool_container/requirements.txt
Flask
google-cloud-storage
gunicorn